requirements.txt
rich
prompt_toolkit
google-genai
python-dotenv
